refactor: 브라우저상에서 'X' 버튼 클릭시 비밀번호를 입력받고 해당 체스게임방 데이터를 삭제

+++ src/main/java/chess/service/RoomService.java
@@ -2,7 +2,9 @@ package chess.service;
 
 import chess.dao.RoomDao;
 import chess.dto.request.RoomRequest;
+import chess.dto.request.UserPasswordRequest;
 import chess.dto.response.RoomResponse;
+import chess.exception.ClientException;
 import java.util.List;
 import org.springframework.stereotype.Service;
 
@@ -24,8 +26,13 @@ public class RoomService {
         return roomDao.findAll();
     }
 
-    public List<RoomResponse> deleteRoomFrom(String id) {
-        roomDao.deleteFrom(id);
-        return roomDao.findAll();
+    public List<RoomResponse> deleteRoomFrom(final String id, final UserPasswordRequest userPasswordRequest) {
+        final String roomPassword = roomDao.findPasswordById(id);
+        final String userInputPassword = userPasswordRequest.getPassword();
+        if (roomPassword.equals(userInputPassword)) {
+            roomDao.deleteFrom(id);
+            return roomDao.findAll();
+        }
+        throw new ClientException("입력한 비밀번호가 일치하지 않습니다! \n다시 입력해주세요.");
     }
 }
